Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > You might want these to be src-relative paths (e.g., //third_party/blink/web_tests/ASANExpectations instead of ASANExpectations), but that's a bit harder to compute. This seems fine in the meantime and might be fine, period.
> 
> Your right I was thinking of a way to do a relative path however its complex. This serves my purpose for now. If I come up with something I will make a CL right away.

It's not actually *that* hard. The tests preserve the general directory structure, so you could find the root of this checkout from this file -- os.path.join(__file__, '../../..') and then compute the relpath of the expectations file (accounting for the fact that the arg might be relative to cwd) from that. 

Not one line of code, but probably not more than 3-4.

Patch-set: 3