Update patch set 7
Patch Set 7: Commit-Queue+2

> Patch Set 7:
> 
> If you're creating a new service with a different runtime, why aren't you using Go instead of python?

That's a great question. There are some technical reasons why, but I completely agree that we should be using Go instead for most of these things in the future. For now:

- Because this repository is being pulled into the Chromium project, we cannot host Go code in the project if we intend it to be used/built in Chromium (because this violates the code policy for Chromium projects).

- Going forward we may be sharing code between this service and the larger catapult project. In particular we may need to re-use the models hosting the test metadata, which is currently not written in Go.

There's even a suggestion that we should be moving the dashboard code out into either a separate repository, or to the infra repository which gives us a lot of other advantages. We haven't gotten to a solid solution there yet, but it's certainly an option.

These are decisions that would be affecting the longer-term development/deployment story for the dashboard, which I must admit we haven't had a solid discussion on yet. We'll probably need to dedicate some time to planning this out more deliberately if we're going to turn the dashboard into a number of smaller services with a mix of different languages for the implementations.

But I asked myself the same question before I got started, and the answer I arrived at was "it would be too much disruption" to not only change the runtime but also the programming language used to implement the solution. :)

Patch-set: 7
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2