Update patch set 6
Patch Set 6: Commit-Queue+1

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > Now we're failing because we don't have `flask` available in the builders:
> > 
> > ```
> > Failed to import test module: sheriff_config.tests.sheriff_config_test
> > Traceback (most recent call last):
> >   File "/b/s/w/ir/cipd_bin_packages/lib/python2.7/unittest/loader.py", line 254, in _find_tests
> >     module = self._get_module_from_name(name)
> >   File "/b/s/w/ir/cipd_bin_packages/lib/python2.7/unittest/loader.py", line 232, in _get_module_from_name
> >     __import__(name)
> >   File "/b/s/w/ir/k/catapult/dashboard/dashboard/sheriff_config/tests/sheriff_config_test.py", line 15, in <module>
> >     from dashboard.sheriff_config import service
> >   File "/b/s/w/ir/k/catapult/dashboard/dashboard/sheriff_config/service/__init__.py", line 12, in <module>
> >     from flask import Flask, request, jsonify
> > ImportError: No module named flask
> > ```
> > 
> > Filed request for vpython wheel to make it available: https://crbug.com/959087
> > 
> > Is there something special we need to do in the build configuration to get the vpython package dependencies installed in the builders?
> 
> This will probably be an issue for whoever tries to deploy the dashboard as well. Dashboard currently gets around that by keeping copies in third_party, if you want to go that route.

This is slightly different with the python37 environment, because App Engine will install the dependencies in a standard location, defined in the `requirements.txt` file. The failure only happens apparently in the way we're setting up the environment on the builder through the LUCI configuration.

Retrying now to see whether the adding of the flask wheel is sufficient, or whether I need to do anything else in the builder configuration to explicitly install `flask` that way.

We already have documentation here on how to get the environment set up so that we can start testing/building the dashboard (and this service) in the `README.md` that comes with the sheriff_config service.

Patch-set: 6
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1