Update patch set 7
Patch Set 7:

Okay, while I try and figure out how to make LUCI builders know about the dependencies we need when testing the service, I'm trying something else out. Here's a few updates:

- I've added a Dockerfile to let isolated testing/build of the sheriff config service.

- I've deployed the current version at https://sheriff-config-dot-chromeperf.appspot.com/service-metadata which seems to work just fine.

- I'm in the process of allowing us to automate the testing and building of the service.

I've updated the instructions on how to build the docker image locally and test. Next steps will be for defining automated "end to end" tests for this service in particular, but we can do that later.

In the process of doing this I've uncovered some issues with the way the imports are defined in the service implementation, which I've been able to fix with the combination of testing the Dockerfile and the trial deployment to the GAE service.

Patch-set: 7
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>