[Typ] Use the same pool for all enabled tests and their retries
Gpu tests start the browser everytime a test set is run, which
also includes the set of tests that are retried. It is unnecessary to
start the browser at the beginning of every retry set, since after
every test failure the Gpu test runner restarts the browser. It also
removes the unnecessary tear down and setup that happens for each retry
test set.

chromium: None
Change-Id: I405129687d9bbc6974069d26840b323a85c294fc