[Trace] Fix flakiness in telemetry_tracing_agent tests
Monkey patching the multiprocessing.Process when the py_trace_event module
is imported, causes Typ to disable tracing control before the
telemetry_tracing_agent tests are run. This causes the
TelemetryTracingAgent.StartAgentTracing to throw an exception. When these
unittests are retried, Typ does not use the multiprocessing module, and
therefore tracing control is not disabled in the retry. This CL will move
the monkey patching into TelemetryTracingAgent.StartAgentTracing after
the trace_enable function
call.

Bug: chromium:960609
Change-Id: I2febf2ec05485ccc566f279107dc5c48ca9586bc