[Trace] Fix flakiness in telemetry_tracing_agent tests
Monkey patching the multiprocessing.Process class in the trace_event_impl,
causes Typ to disable tracing control before the telemetry_tracing_agent
tests are run. This causes the TelemetryTracingAgent.StartAgentTracing
to throw an exception. When these unittests are retried, Typ does not
use the multiprocessing module, and therefore tracing control is not
disabled in the retry.

Bug: None
Change-Id: I2febf2ec05485ccc566f279107dc5c48ca9586bc