Reland "[Telemetry] Make desktop retries reset environment."
This is a reland of d6c7b00c216b44f56429ee875f960ca4fa740c37

The reason this caused failures was because I used python's mock
library to replace browser.Browser with a mock, and then I didn't
replace it with the real version at the end.

patchset 1 is the original change, and patchset 2 has my fix.


Original change's description:
> [Telemetry] Make desktop retries reset environment.
>
> Bug: chromium:944343
> Change-Id: Ide68966dd0c4008e00299749c7f2970b88dd6297
> Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1541442
> Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
> Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
> Reviewed-by: John Chen <johnchen@chromium.org>
> Reviewed-by: Kenneth Russell <kbr@chromium.org>

Bug: chromium:944343
Change-Id: Iba25fa1936d932dfc77e4b4352a6b3a65917c734