Pass Blink image decode TRACE_EVENT timing to rendering_metrics.
This change surfaces the CPU duration of calls to
ImageFrameGenerator::decode[ToYuv] for RGB or YUV decoding,
respectively. These are currently trace events in Blink. New image
decoding Telemetry stories landing in crrev.com/c/1536460 motivate this.

It also records the CPU duration of
GpuImageDecodeCache::DecodeImageIfNecessary, which is a trace event in
category cc and is a common path for YUV and RGB decoding in GPU
rasterization. This metric should improve when the YUV decoding feature
tracked at crbug.com/900264 is turned on, whereas the other two metrics
need to be superimposed to show the improvement (since a single decode
is either RGB or YUV).

Change-Id: I6167b641dbdb5bacd31448886f51d319cf58eb08
Bug: chromium:944756