Update patch set 6
Patch Set 6:

> Patch Set 4: Code-Review+1
> 
> Thank you. The code looks good to me; I just have one question about the new histograms.
> 
> You are adding the duration of each decodeToYUV call as a data point to the histogram. Are the image chunks that are passed to decodeToYUV supposed to be of the same size? I mean if a future change breaks down chunks sizes in half (i.e. two func calls on 500K chunks instead of one func call on a 1M chunk) you will see a drop in you metric, although the efficiency of decodeToYUV is not changed. If that shouldn't happen maybe you should compute the sum of all durations? Or maybe you should log the chunk sizes, too, in the trace event and look at something like cpu duration per 1M chunks?
> 
> I'm not familiar with that part of the code. Just wanted to make sure you have considered this.

Yes, that is something I'm aware of. For the test that's motivating these metrics, we don't go through the incremental decoding path (and should hopefully avoid it anyway because the image data is directly encoded in base64 in src...though I'm not familiar with resource loader code). Furthermore, it's documented for each metric that the timing is per image data chunk (unevenly sized). I agree it'd be useful to have it on a per-image basis but I think that requires more thinking/sophistication to hit all cases of images and decoding.

Patch-set: 6
Reviewer: Gerrit User 1301626 <1301626@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>