Remove unused chrome_v8.json testdata file
This patch removes an unused testdata file that contains many references
to the 'chrome-devtools://' url scheme.

The 'chrome-devtools' scheme is being replaced with 'devtools://' as
seen in patch 1582881:
https://chromium-review.googlesource.com/c/chromium/src/+/1582881

You can see in the discussion of the patch above that it was recommended
to try deleting this file as it is no longer used.

Change-Id: Iedf9b64858173e9ad8243f97dd3430fa8d954aea