[Trace] Fix flakiness in telemetry_tracing_agent tests
When the py_trace_event.trace_event module is imported, the
multiprocessing.Process class is monkey patched with a wrapper for the
multiprocessing.Process class. Typ then uses this wrapper class to run the
unit tests. The wrapper disables tracing control for each child process it
spawns. telemetry tracing agent unit tests see that tracing control is
disabled and then raises an exception. When Typ retries the telemetry tracing
agent unit tests it does not use the multiprocessing.Process class therefore
does not disable tracing control. Then no exception is raised in the retries.
This CL will move the monkey patching to the the StartAgentTracing function
so that the wrapper class can only be used after calling StartAgentTracing.

Bug: chromium:960609
Change-Id: I2febf2ec05485ccc566f279107dc5c48ca9586bc