Update patch set 43
Patch Set 43:

> Patch Set 41:
> 
> > Patch Set 41:
> > 
> > > Patch Set 40:
> > > 
> > > > Patch Set 40:
> > > > 
> > > > I used the dev link you provided, loaded a few alerts, they all seem to show 2 bisect buttons? Why is it doing that?
> > > 
> > > In the V1 UI, users can only click a single point, but v2spa allows users to brush over multiple ranges of points. There may be zero, one, or many alerts within each range. One old feature request was to start bisects over a range of points more easily. But users should also be able to start bisects for alerts. So the first Bisect button is for the alert, and the second Bisect button is just for the revision range without considering the alert.
> > > I changed the labels of the buttons to 'Bisect Alert' and 'Bisect Range'.
> > > I'm also considering hiding the 'Bisect Range' button in cases where it would do the same thing as the 'Bisect Alert' button.
> > > WDYT?
> > 
> > Hmm I find the double button confusing, why not just have 1 that's the range selected? Presumably if you click on the alert, the range narrows to the alert anyway?
> 
> ^ Good Q. Is there a use case that requires separate bisect buttons? It does seem confusing (even if the labels are clearer now).  Users will still have to stop and think about which one to click, but can the UI figure out which action makes sense and just use a single button for both?

I removed the bisect-dialog from alert-detail, leaving only the one in details-table.
I added the revision range to the bisect button label.
I added a bisectMessage to ask the user to sign in. I noticed that I'd forgotten to make a Polymer 3 version of chops-signin-aware, so I added that here, with a todo to remove it when spa migrates to lit-element, when it can use the version in npm.

Since the bisect button is unavailable when multiple suites/bots/cases are aggregated into one lineDescriptor/body, I added a button to open an alert's timeseries (1 suite, 1 bot, 0-1 case[s]) in a new chart-section.

PTAL

Patch-set: 43
Reviewer: Gerrit User 1115874 <1115874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>