[py_utils] Add exc_util.BestEffort decorator
The decorator allows to log and dismiss exceptions on a closing/cleanup
function, if an existing exception is already being handled.

Otherwise, without the decorator, the second exception is raised while
the original one, usually more important is lost.

If the decorated function is called in a context where no exceptions
are being handles, then exceptions raised by the function are passed
on to callers as usual for them to handle.

Bug: chromium:958778
Change-Id: I611e6945f278b75b1d4f07cab0896943e67cba1a