[Typ] Use a file to store sub process output
When Typ recursively run's itself, the output goes to the terminal. This
can make the final output very verbose and almost unreadable. If we can
catch this output and make it part of the calling tests output when it
fails, then we can drastically reduce the size of the final output.

Bug: chromium:960833
Bug: chromium:960606
Bug: chromium:952105
Change-Id: I1943dbd65e44c7bcb6f465620905e92398eb45a0