Update patch set 12
Patch Set 12:

(4 comments)

I think I'm still not understanding quite what's going on and why you're making the changes you are.

I think I get at least part of it, which is that before this fix, if we call host.capture_output() prior to running a test, then anything in the python code that calls print() will get captured, but if something calls subprocess.call() directly that will still have access to (and use) the underlying streams, and hence the output is leaking through?

And, in order to fix this, subprocess needs actual file descriptors to pass to the subprocess? (and so we need to change _TeedStream to use a real file instead of an in-memory thing)?

Is that right? If so, I think I'm starting to understand, but I do have a few more questions (as per the individal comments).

Patch-set: 12
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>