SheriffRedux: Implement a luci-config client
This uses python37 environment-safe utilities for making requests to the
luci-config service, for sheriff configuration files. This change does
the following as well:

- Define a cron job which will regularly poll the luci-config service.

- Store the configuration data by indexing the individual sheriff
  configuration instances to the project config and revision.

- Exposes an API for finding all the sheriff configurations which match
  a given timeseries identifier.

We're isolating the service which matches the sheriff configurations to
the timeseries identifiers to allow us to consolidate the implementation
of the logic which determines the sheriff configurations that match.

R=abennetts@chromium.org, nodir@chromium.org, simonhatch@chromium.org

Bug: chromium:953181
Change-Id: I87da4679a489da0b34782a98f461954353b389cc