Pinpoint: Fix comparison to handle None sentinel
This change ensures that we can apply the Comparison function and handle
the first/sentinel case where the left-hand-side of the reduction is a
None. This is a result of removing references to xrange() which has been
removed in Python3, and in the conversion we learned that this code
wasn't covered by specific testing.

R=perezju@chromium.org, simonhatch@chromium.org

Bug: chromium:966328
Change-Id: I199fe057e99785b83477f1bbde1df39d26d532ca