[FCP++] Filter loading image and text candidate
In the context of FCP++, the image/text candidates that are "loading"
should block being reported to UKM/UMA. This design is in order to
reduce the abort bias of the metrics. The reason is simple. From
user's perspective, when they find the largest content still loading,
they expect the main content to finish loading in the future, and they
would hesitant to call the already-loaded largest content as the most
important element.

In the trace, these loading candidates are named as
"LargestImagePaint::NoCandidate" or "LargestTextPaint::NoCandidate"
after we landed CL[1].
We should block telemetry from reporting any candidate if they are
the latest candidates. That way the telemetry metric will be aligned
with the UMA/UKM candidate.

[1] crrev.com/c/1610566:
[FCP++] Text and Image should reset candidate in trace when necessary

Bug: chromium:960365
Change-Id: I5208380cc0739bd6f937221f39d069148bcf8927