[Telemetry] Factor out code for _ShouldRunBenchmark check
This makes the logic of the method easier to follow, and simplifies the
(already quite complicated) RunBenchmark function.

Also take the chance to simplify the code under the assumption that
there are no browserless benchmarks any more.

And fix benchmark_unittest so it uses an "options" object including the
necessary story_runner options that RunBenchmark expects.

Bug: chromium:866458
Bug: chromium:840428
Change-Id: Icef665061b70ed62bcc2a00e4d37defad61cf163