[Typ] Store relative paths for expectations files paths in metadata field
Previously we changed the list of expectations file paths to a list of
file names. However this does not give much information on where the
test expectations paths are stored. This CL will attempt to create a
list of relative paths to test expectations files using the top_level_dirs
list in the Runner instance.

Bug: chromium:698902
Change-Id: I31324eef8a903be2ef1d6f8c31826da0134e26f4