[Typ] Store relative paths for expectations files paths in metadata field
Previously we changed the list of expectations file paths to a list of
file names. Previously it was a list of absolute file paths. We did this
because the results merger script was failing because the absolute paths
of test expectations files were different on some bots. The file name does
not give much information on where the test expectations paths are stored.
This CL will attempt to create a list of relative paths to test expectations
files which will be the same across all bots using the top_level_dirs list.

Bug: chromium:698902
Change-Id: I31324eef8a903be2ef1d6f8c31826da0134e26f4