Update patch set 10
Patch Set 10:

This is making the paths relative to the top_level_dir, and I'm not sure how clear that will be for people, partially because I'm not sure how often people will know what the top_level_dir for a test suite is or whether the expectation file will live relative to the top_level_dir.

Where do the expectations files live for the GPU tests, and what is the top_level_dir for them?

In https://chromium-review.googlesource.com/c/catapult/+/1592693/3#message-abad9628d9281d476dda7135693efb6ca6f52bd3, I was actually suggesting that you use src-relative paths, which are different. However, in re-thinking about this now, I realize that it's not actually obvious what the src to use is. For example, when running tests in chromium, you want to use paths relative to src/, but typ lives in the catapult repo, which is DEPS'ed in, so it's four levels down. However, if you run the tests in catapult, you're only two levels down. You can't just crawl up to the directory that contains the .git dir, because that won't be present in the isolate.

Maybe we need to modify the test runner to pass in the directory to make things relative to use, and standardize them on passing in the top of the relevant repo.

Patch-set: 10
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>