Add a new total_rendering metric
total_all_cpu_time_per_frame is noisy on Windows. The new metric will
exclude some threads that are causing noise but include threads that
do graphics work so that CPU regressions caused by the graphics team
can be detected more easily. We may or may not need to keep the new
metric if the noise in total_all_cpu_time_per_frame on Windows is
addressed. Please see the bug for more details.

Bug: chromium:964423
Change-Id: I34a6aee258115e8d49916474093873613bd7388a