Fix exception in run_chromeos_tests.py caused by crrev.com/c/1611755
Changes in the above CL require that the fileno method is overriden
for any stream class that is used for the Host's stdout and stderr
member variables. This is so that any sub process's output can be captured
when host.capture_output is called. This CL will override the
method for the _LoggingOutputStream object and return sys.stdout's file
descriptor. Any output from a sub process spawned by a test will go
to standard out. Also the Typ.Host instance can capture this output.

Bug: chromium:967922
Change-Id: I5cd82171bb154de00e7d78275105d642fb51db28