Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

I agree that TBR is not a great solution. I think we should either move the files which Dean needs to frequently change to a new directory and make him owner of that or simply leave the files where they are today and make him owner of catapult. The drawback seems to be that Dean may start getting more code review requests (for things outside of catpult/dashboard) but he could always forward such code reviews to other folks. We should try to accommodate folks that are working in different timezones.

Patch-set: 1