SheriffRedux: Move skip_files to .gcloudignore
The `gcloud app deploy` script now errors out when we have both. This
change moves the skip patterns into .gcloudignore instead.

We also remove the Dockerfile and cloudbuild.yaml configuration since
App Engine Flexible doesn't let us define these files if we want to use
the flexible Python environment.

We also learn that we've been using the `google.auth` package
incorrectly, until we tried deploying a version of the app, which we're
fixing too.

Since we haven't set up Google Cloud Build to perform post-commit
end-to-end testing, local testing reveals some issues with the tests
that we're also updating in this change.

R=orodley@chromium.org

Change-Id: I4403723dc2881cd9daa5f8f8fedf1310e5e963d4