[Telemetry] Fix reporting of tracing errors during DumpMemory
A previous CL tried to make TracingUnrecoverableException
a subclass of AppCrashException in order to obtain further debug
information (e.g. logs and possible crash stacks) for those errors.

This however did not work as AppCrashException needs a reference
to the "app" (browser) that crashed; and the app object is not
available at that level of abstraction (the tracing_backend).

In this CL we fix the previous mistake and, in the call to
DumpMemory, re-raise any TracingUnrecoverableException's as
AppCrashException's to help debug the issue in the linked bug.

Bug: chromium:964960
Change-Id: If76e831d1b49c811e151ba8d1b5c1310816d86a5