Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > > Patch Set 4:
> > > > > 
> > > > > (1 comment)
> > > > > 
> > > > > What would actually get reported in the links?
> > > > 
> > > > The particular use case in mind is triage links for Gold tests so that they're easier to get to. Currently, they require digging through the task output logs to get. https://bugs.chromium.org/p/chromium/issues/detail?id=968276
> > > 
> > > So this might be a link per test failure or something like that, rather than a link per overall test run?
> > > 
> > > This is why I want to be a little careful about how we spec this, I don't necessarily want this to be something where we send arbitrary JSON along, but we should be clear about which things might be repeated in a more semi-structured way, rather than just a single key-value list.
> > 
> > Yes, there would be a link per test failure (or at least per test failure that ends up generating a triage link).
> 
> I should note that the only reason we're going with this particular approach for surfacing this data is that the merge script documentation seemed to imply that the "links" field was already at least somewhat standardized "[JSON results] may optionally contain a top level "links" field that may contain a dict mapping link text to URLs, for a set of links that will be included in the buildbot output."
> 
> It's starting to sound like that's not actually the case, though.

Yes, I think that's right. 

Looking at crbug.com/968276, I see two different examples. The first is kbr@'s, which is a link to a specific log file. I think the way to handle that is in the per-test 'artifact' dict. I think the Skia Gold link is probably a slightly different variant of that.

The recipe should probably then have generic logic for providing links to the artifacts for the failing tests.

I don't think we want the harness itself trying to decide which things are important and which aren't, because then we end up with really inconsistent output in the UI (which is what we have today, except that the inconsistency comes from the different step classes in the recipe, not from the harness).

How does that sound?

Patch-set: 4
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>