SheriffRedux: Use google-auth when polling
Because we're using Google Cloud Endpoints for the sheriff-config
service, we need to use a modern google-auth implementation. Previous
CLs introduced the dependencies we need being vendored into the project,
which we're bringing together in this CL.

SheriffRedux: Vendor requests-toolbelt

This is a requirement of the google-auth library, which we are also
vendoring until we can move to Python3 and App Engine's Python3 Standard
Evironment, to rely on pip dependencies.

R=simonhatch@chromium.org

Bug: chromium:953181
Change-Id: Ie6d18fb3122c29375c56f8f7d0a46a290c731279