SheriffRedux: Handle empty configs cases
This fixes a case where we're polling luci-config but getting an empty
result set (which is valid). Here, instead of raising an exception, we
instead treat an empty input as a signal that we should update the index
to have no configs.

We're also treating the case where the input is None as a valid input.

R=orodley@chromium.org, simonhatch@chromium.org

Bug: chromium:953181,chromium:971170
Change-Id: I7e0b1964d7cca021aa0a83dcad10921259909fe7