[metrics] Optimize V8 and Blink GC metrics
Currently these metrics iterate over all slices in the model to the
relevant GC events. This patch narrows the search to only renderer
process and only to the threads that can execute JavaScript code
(the main thread of a renderer and web worker threads).

GC metrics computation time reduces by 97%.

Bug: chromium:958441
Change-Id: Ie86d31678e5358dcae48b01882cc9835c7c35925