[Telemetry] Always store artifacts if there is an output_dir
We already store artifacts for all output formats except "none" and
"gtest". In the future, where we expect traces (and thus artifacts)
become the only output of Telemetry; not collecting artifacts at
all becomes less and less reasonable.

With this change we make it so that artifacts are always collected
as long as an output_dir is specified (which is also almost always
the case, except for some unit tests).

Bug: chromium:972025
Change-Id: I370fb18a7e5a81c23c09f989abf802235b9ba82f