[typ] Add typ implementation of artifacts
Adds an ArtifactResults implementation to typ that's similar to the one
provided by Telemetry. This is so that tests that use typ for generating
the results JSON file instead of Telemetry's implementation can still
use artifacts.

The two implementations may be able to be joined later on once some
refactoring work on Telemetry's end is finished.

Bug: chromium:968276
Change-Id: I4624d978caeee356f869a4db8e692139acd3f805