Pinpoint - Handle change_id's that resolve to multiple changes.
Some changes that are merged to other branches can resolve to multiple
ChangeInfo's in gerrit, with information about which branches and when
they were merged in. When you use the change_id without specifying both
project and branch, ie. chromium/src~master~<change_id> the change_id
can resolve to multiple changes.

We can use the patch module to grab both the
repo path and branch, specifying the change id more exactly so that it
resolves to only 1.

Bug: chromium:951779
Change-Id: I5830555ca4e6da7792ccefdca53f88be9a0a0a14