[Telemetry] Pass --repository-absolute-path to Typ
The absolute path of the source directory will be passed as the value of
this command line argument. Typ will then use this value to derive relative
paths for files. One such path is the expectations file path. The relative
path of the expectations file will show in the full JSON results, which can
give a better idea of where the files are located in the chromium repository.

Bug: chromium:698902
Change-Id: I19c291e4a1c970a4a6141223d8c420de4a3eb905