[Typ] Remove useless code that was used to skip tests
Before the new classification API was implemented, a disabled test would be
sent to the job pool to run. _run_on_test would check if the test had the skip
expectation. If it did then it would be skipped. After the classification API
was implemented, this now happens when tests are classified. Therefore a disabled
test would never be sent to a job pool.

Bug: chromium:945481
Change-Id: Ib77112f1932566023334cb0b8afe05018a181944