[Telemetry] Expose option to remove system packages
Adds RemoveSystemPackages() to android_platform.py and
android_platform_backend.py, which simply call into the existing method
in system_app.py.

This is necessary in order to run the VR
perf tests on the swarmed Pixel 2 devices, which have VrCore installed
as a system app, preventing us from installing a specific version that
is used for testing until the system app is removed.

Bug: chromium:972221
Change-Id: I0d842f2356d72de9cd00228d751bf15e1d414314