Fix media video playback metrics -- take 2
PS#1 is the original change, later patch sets are the fixed one.

------------ Original Description ------------
We (sometimes) moved off of the compositor thread for video metrics when
VideoSurfaceLayer was turned on by default in M75 by
4f95a06f1b6612789b910f6b0eaa33496267efba.

Unfortunately the thread name is a bit ugly, so we end up looking
through all the base::PostTask threads to find the video metrics.

BUG=chromium:965682, chromium:971441

Change-Id: I11958973bab4fb7812915cd65c8002c4492dbbfb