Update patch set 9
Patch Set 9:

(1 comment)

Thank you for looking so closely!

> Patch Set 9:
> 
> (2 comments)
> 
> Looking good - Some high level comments before diving too deeply in: 
> 
> - Optional thought: For keyboard interactions, I've often found it useful to leverage the browser's native focus features to save some of the work of writing keyboard code. For example, in this case, if you wanted, rows could be buttons with click handlers that can be natively tabbed through then additional key bindings (ie: arrow keys) could be implemented as an additional layer on top that inter-ops. Not sure if this makes sense for the specific architecture you're going for here, especially since saving the cursor in the state seems to be a major part of this implementation, but thought it might be useful to mention. 

If the focus switched from the input to the options, then users would not be able to resume typing a query without tabbing back to the input.
Also, if each option was tabbable, then users would need to tab through all of them to get to the next menu-input. Also, they wouldn't be able to get from the bottom of the options back to the top easily.

> - Keyboard usability nit: Once I focus an input with these keystrokes bound, it seems that I can no longer tab to escape the input. I can use Esc to exit, so I assume the content is meant to be key trapping. Shift+Tab does work though, which feels a bit confusing. While using tab, having the focused trapped feels a bit like a break from the flow of keyboard navigating through the page. I think this might flow a bit better if you could tab between the menu elements as well, even if the focus is still trapped, so you don't feel "stuck". 

It looks like I can tab through menu-inputs in the chart-section's timeseries-descriptor, but it gets stuck in the Statistic menu-input, and in the report-section's menu-input.
It looks like you can get back to tabbing through if you hit Escape to blur the menu-input.
I might like to leave polishing these nits for when inspiration strikes at some later date if that's ok.

> - Since the left and right arrow are caught as well, is there any concern about having the left and right arrow key handlers collide with the case when the user uses left and right to move their cursor within the input?  

If the user has typed a query, then left/right arrow keys toggle the isExpanded boolean, but don't affect which options are visible. If the user has typed a query, then they're looking for something, which is a different goal than showing/hiding branches, so they should add more query terms, not try to use the left arrow to hide branches.

> - I'm not able to reproduce it reliably, but I think there might be a few bugs in this code - after playing around some, sometimes when I focus the input, the cursor doesn't appear anymore in the dropdown menu, and I don't seem to be able to navigate between elements.  There also seems to be a bug (I think with nested options) where the cursor becomes invisible at times (I think it's trying to select options that are collapsed). 

I'm sure there are bugs. It's a surprisingly complex feature. But I've tested it reasonably well and tried to design it to support tweaks to fix bugs, so I feel it's at the "perfect is the enemy of the good" stage. Always happy to hunt down specific issues with repro steps.

> - This might be unrelated to this change, but I noticed that when I tried typing in the input a lot and deleting some text seemed to cause the Perf Dashboard to freeze in my browser.

If the query and option tree are complex, then arrow key presses could possibly get to be fairly CPU intensive, but I haven't noticed any jank on my macbook. Want to try capturing a devtools trace?

Patch-set: 9