[catapult] Workaround for "Tracing Dev Server" hangs
Trying to run tracing/tests was often failing due to modules not being
fully loaded before the test runner started trying to load/run tests.

This CL adds a wait between the time when HTMLImports were supposed to
be ready, and the start of the test runner itself.

Also fix an inssue where the dev_server would not stop in case of
errors while loading the test framework; and add a bit of extra debug
information on run_dev_server_tests.py.

Bug: chromium:973847
Change-Id: I98e96f4efed1fd1eb54c0b671c5144189c551c69