[Devil] Restore root after system app modification
Changes the EnableSystemAppModification context manager to restore the
root state of the device after its final reboot. This is due to some
parts of Telemetry (and possibly elsewhere) seeming to implicitly
assume that root has already been granted, which is most likely the
case, but will not be immediately after a reboot.

Bug: chromium:972221
Change-Id: Iee77f2d47f1392e62342d2cd6b6c39b79748d88b