[Telemetry] Add new methods to get tags for Typ's test expectation format
Adds the GetTypExpectationsTags to telemetry's PossibleBrowser, Platform and
PlatformBackend classes. These functions will be used to return a list of tags
representing the OS name, OS version name and browser type. These tags will be
passed to Typ's expectation_parser module so that they can be used to find
active test expectations in a test run.

Bug: chromium:973936
Change-Id: Ic84c00ffaa0e5d50c187a7f686e266bda9b60966