[Telemetry] Add new methods to get tags for Typ's test expectation format
Adds the GetTypExpectationsTags to telemetry's PossibleBrowser, Platform, Browser
PlatformBackend, MacBackend and AndroidBackend classes. These functions will be used to return
at the very lease a list of tags representing the OS name, OS version name and browser type
for all platform back ends. Android and Mac back ends will be treated as special cases. For
the Android back end it will return additional tags representing the model name the android
and if it has a svelte (low memory) build. For the Mac back end it will return additional
tags representing the Operating system version number if the version number is either
10.11 or 10.12. These tags will be passed to Typ's expectation_parser module so that they
can be used to find active test expectations in a story run.

Bug: chromium:973936
Change-Id: Ic84c00ffaa0e5d50c187a7f686e266bda9b60966