Pinpoint - Move all job and quest exceptions to common spot.
This moves all the exceptions that pop up during the running of a job
into a common spot. Goal is to go through and reword them to be
clearer, as well as modify the UI to only show the readable portion and
hide the exception stack info until bug filing.

Bug: chromium:971370
Change-Id: Ifd700587688a0a2ab4b9735ed92a04b3d7dc6389