[Telemetry] Add back a grouping_label argument for some Values
A previous CL [1] replaced 'tir_label' with a new 'grouping_label' used
to group labels during chart json summarization.

Unlike 'tir_label', the new 'grouping_label' is computed statically
from the page object passed to the Value constructor. Which worked well
to replicate the old grouping behavior in most circumstances.

However, this did not work for summary values (which do not have a page)
and instead the old code relied on the value 'merge' functions
copying the 'grouping_label' (if same for all value merged) into the
new merged value.

This CL adds back that behavior.

[1]: https://crrev.com/c/1651469

Bug: chromium:976257
Change-Id: Ib6829ba8ff437774d350b5cbbe89d5a8b2fb55e3