Clarify the error message for Trace View picture debugging
When using remote tracing, i.e. getting a trace from an Android
device and viewing it on desktop, it's not obvious that the desktop
browser is the one that needs the --enable-skia-benchmarking flag,
users (such as me) may be trying to set it on the instrumented device
and then wonder why it's not working.

Clarify the error message to refer to the browser displaying the
trace specifically.

Change-Id: I26037124684037832c5dba5cfd6a023997db52fe